Translate redis kernel space commandtypes to user-space commandtypes
